feat(junit-playwright): breakup extension into separate objects for readability
