+++ playwright-junit/src/main/java/com/microsoft/playwright/PlaywrightExtension.java
@@ -1,31 +1,26 @@
 package com.microsoft.playwright;
 
-import org.junit.jupiter.api.extension.*;
-import org.junit.jupiter.api.parallel.ExecutionMode;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.jupiter.api.extension.ParameterContext;
+import org.junit.jupiter.api.extension.ParameterResolutionException;
+import org.junit.jupiter.api.extension.ParameterResolver;
 
-import java.util.HashMap;
-import java.util.Map;
+import static com.microsoft.playwright.ExtensionUtils.hasProperAnnotation;
 
-import static com.microsoft.playwright.JUnitUtils.getExecutionMode;
-import static org.junit.platform.commons.support.AnnotationSupport.findAnnotation;
-import static org.junit.platform.commons.support.AnnotationSupport.isAnnotated;
-
-public class PlaywrightExtension implements ParameterResolver, AfterAllCallback, AfterEachCallback {
+public class PlaywrightExtension implements ParameterResolver {
   private static final ThreadLocal<Playwright> playwrightThreadLocal;
-  private static final ThreadLocal<Map<Class<? extends BrowserFactory>, Browser>> threadLocalBrowserMap;
 
   static {
     playwrightThreadLocal = new ThreadLocal<>();
-    threadLocalBrowserMap = ThreadLocal.withInitial(HashMap::new);
   }
 
   @Override
   public boolean supportsParameter(ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
-    if(!hasProperAnnotation(extensionContext)) {
+    if (!hasProperAnnotation(extensionContext)) {
       return false;
     }
     Class<?> clazz = parameterContext.getParameter().getType();
-    return Playwright.class.equals(clazz) || Browser.class.equals(clazz);
+    return Playwright.class.equals(clazz);
   }
 
   @Override
@@ -34,31 +29,7 @@ public class PlaywrightExtension implements ParameterResolver, AfterAllCallback,
     if (Playwright.class.equals(clazz)) {
       return getOrCreatePlaywright();
     }
-
-    if (Browser.class.equals(clazz)) {
-      BrowserFactory factory = getBrowserFactoryInstance(extensionContext);
-      return getOrCreateBrowser(factory);
-    }
-
     throw new ParameterResolutionException("Unable to resolve Playwright-related parameter");
-  }
-
-  @Override
-  public void afterAll(ExtensionContext extensionContext) {
-    if (getExecutionMode(extensionContext) == ExecutionMode.SAME_THREAD) {
-      threadLocalBrowserMap.get().keySet().forEach(PlaywrightExtension::closeBrowser);
-      threadLocalBrowserMap.remove();
-      closePlaywright();
-    }
-  }
-
-  @Override
-  public void afterEach(ExtensionContext extensionContext) {
-    if (getExecutionMode(extensionContext) == ExecutionMode.CONCURRENT) {
-      threadLocalBrowserMap.get().keySet().forEach(PlaywrightExtension::closeBrowser);
-      threadLocalBrowserMap.remove();
-      closePlaywright();
-    }
   }
 
   public static Playwright getOrCreatePlaywright() {
@@ -78,50 +49,6 @@ public class PlaywrightExtension implements ParameterResolver, AfterAllCallback,
       System.out.println("Closing Playwright " + getPlaywright());
       getPlaywright().close();
       playwrightThreadLocal.remove();
-    }
-  }
-
-  public static void closeBrowser(Class<? extends BrowserFactory> browserFactoryClass) {
-    Browser browser = getBrowser(browserFactoryClass);
-    if (browser != null) {
-      System.out.println("Closing Browser " + browser);
-      browser.close();
-    }
-  }
-
-  public static Browser getOrCreateBrowser(BrowserFactory factory) {
-    Class<? extends BrowserFactory> browserFactoryClass = factory.getClass();
-
-    if (getBrowser(browserFactoryClass) != null) {
-      return getBrowser(factory.getClass());
-    }
-
-    Playwright playwright = getOrCreatePlaywright();
-    Browser browser = factory.newBrowser(playwright);
-    System.out.println("Created Browser " + browser);
-    threadLocalBrowserMap.get().put(browserFactoryClass, browser);
-    return browser;
-  }
-
-  public static Browser getBrowser(Class<? extends BrowserFactory> browserFactoryClass) {
-    return threadLocalBrowserMap.get().get(browserFactoryClass);
-  }
-
-  private boolean hasProperAnnotation(ExtensionContext extensionContext) {
-    return isAnnotated(extensionContext.getTestMethod(), UseBrowserFactory.class) ||
-      isAnnotated(extensionContext.getTestClass(), UseBrowserFactory.class);
-  }
-
-  private static BrowserFactory getBrowserFactoryInstance(ExtensionContext extensionContext) {
-    UseBrowserFactory useBrowserFactory =
-      findAnnotation(extensionContext.getTestMethod(), UseBrowserFactory.class)
-        .orElse(findAnnotation(extensionContext.getTestClass(), UseBrowserFactory.class)
-          .orElseThrow(() -> new PlaywrightException("UseBrowserFactory annotation not found.")));
-
-    try {
-      return useBrowserFactory.value().newInstance();
-    } catch (InstantiationException | IllegalAccessException e) {
-      throw new PlaywrightException("Unable to create an instance of the supplied BrowserFactory", e);
     }
   }
 }
