+++ playwright-junit/src/test/java/com/microsoft/playwright/BrowserSameThreadTests.java
@@ -6,7 +6,8 @@ import org.junit.jupiter.api.parallel.ExecutionMode;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@TestMethodOrder(MethodOrderer.MethodName.class) // this is here to force customBrowserTest1 to run before customBrowserTest2
+@TestMethodOrder(MethodOrderer.MethodName.class)
+// this is here to force customBrowserTest1 to run before customBrowserTest2
 @Execution(ExecutionMode.SAME_THREAD)
 @UseBrowserFactory
 public class BrowserSameThreadTests {
@@ -61,7 +62,6 @@ public class BrowserSameThreadTests {
     assertEquals("firefox", browser.getName());
     assertEquals(customBrowser, browser, "Custom browser is not the same as the one created in the first test that uses this browser");
   }
-
 
 
   @Test
