+++ playwright-junit/src/test/java/com/microsoft/playwright/PlaywrightConcurrentTests.java
@@ -12,6 +12,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 @UseBrowserFactory
 public class PlaywrightConcurrentTests {
   private Playwright p;
+
   @BeforeEach
   void beforeEach(Playwright playwright) {
     assert playwright != null;
